Create bootstrap server
