+++ package.json
@@ -9,6 +9,7 @@
     "dev": "ts-node-dev src/server"
   },
   "dependencies": {
+    "chalk": "^4.1.0",
     "dotenv": "^8.2.0",
     "express": "^4.17.1",
     "pg": "^8.2.1",
